OpenAI Integration complete
